static_root/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.6
cssselect2==0.4.1
Django==3.2
django-appconf==1.0.5
django-compressor==2.4.1
django-environ==0.8.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.3
django-ordered-model==3.5
django-rosetta==0.9.7
django-summernote==0.8.11.6
djangorestframework==3.12.4
easy-thumbnails==2.8
Faker==4.18.0
idna==3.2
lxml==4.6.4
Pillow==8.3.2
polib==1.1.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
rcssmin==1.0.6
reportlab==3.6.3
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.1
svglib==1.1.0
text-unidecode==1.3
tinycss2==1.1.1
urllib3==1.26.7
webencodings==0.5.1
